+++ packages/components/modules/messages/index.ts
@@ -1,6 +1,3 @@
-export { default as MessageItem } from './MessageItem'
-export type * from './MessageItem/types'
-
 export { default as MessagesList } from './MessagesList'
 export type * from './MessagesList/types'
 
@@ -15,6 +12,7 @@ export type * from './SendMessage/types'
 
 export { default as CreateChatRoomList } from './CreateChatRoomList'
 export type * from './CreateChatRoomList/types'
+
 export * from './context'
 
 export * from './graphql/mutations/SendMessage'