+++ packages/components/modules/profiles/graphql/queries/ProfileItem.ts
@@ -2,7 +2,6 @@ import { graphql } from 'react-relay'
 
 export const ProfileItemFragment = graphql`
   fragment ProfileItemFragment on Profile
-  @inline
   @argumentDefinitions(avatarSize: { type: "Int", defaultValue: 100 }) {
     id
     name